requirements.txt
python-dotenv==1.0.0
pinecone-client==2.2.4
langchain-core
langchain
langchain-community==0.0.13
langchain-huggingface==0.0.3
pypdf==4.0.2
sentence-transformers 